renamed visualize functions to better reflect their purpose; moved CLIOutput to its own file
